+++ Backend/Domain/Common/Entity.cs
@@ -1,8 +1,6 @@
-﻿
-
-namespace Domain.Common
+﻿namespace Domain.Common
 {
-    public abstract class  Entity
+    public abstract class Entity
     {
         public int Id { get; protected set; }
     }
