+++ Backend/Domain/Entities/ArchivoListaReproduccion.cs
@@ -1,23 +1,23 @@
-﻿using Domain.Common;
+﻿using Domain.AggregateRoots;
+using Domain.Common;
 
 namespace Domain.Entities
 {
-   public class ArchivoListaReproduccion : AggregateRoot
+    public class ArchivoListaReproduccion : AggregateRoot
     {
         public int IdLista { get; private set; }
         public int IdArchivo { get; private set; }
 
-        public ListaDeReproduccion ListaDeReproduccion { get; private set; }
+        // Navegación
+        public ListaDeReproduccion ListaDeReproduccion { get; set; }
         public Archivo Archivo { get; set; }
+
         protected ArchivoListaReproduccion() { }
 
-        public ArchivoListaReproduccion(int idLista, int idArchivo)           
+        public ArchivoListaReproduccion(int idLista, int idArchivo)
         {
             IdLista = idLista;
             IdArchivo = idArchivo;
         }
     }
-
-    
-
 }
