+++ Backend/Domain/Entities/Credito.cs
@@ -1,18 +1,20 @@
-﻿using Domain.Common;
+﻿using Domain.AggregateRoots;
+using Domain.Common;
 
-namespace Domain.Entities
+namespace Domain.Entities // o AggregateRoots si es necesario
 {
-    public class Credito : AggregateRoot
+    public class Credito : Entity
     {
+        public int Id { get; private set; }
         public int IdUsuario { get; private set; }
         public int Cantidad { get; private set; }
         public DateTime FechaAdquisicion { get; private set; }
 
-        public Usuario Usuario { get; set; }
+        public Usuario Usuario { get; set; }  // Esta propiedad debe existir
 
         protected Credito() { }
 
-        public Credito (int idUsuario, int cantidad)
+        public Credito(int idUsuario, int cantidad)
         {
             IdUsuario = idUsuario;
             Cantidad = cantidad;
