+++ Backend/Domain/Entities/Factura.cs
@@ -1,3 +1,4 @@
+using Domain.AggregateRoots;
 using Domain.Common;
 using System;
